refactor: move the semantic-check logic to AlloySolver
